refactor : ItemChecker에서 BidDetectService.kt를 분리
