+++ src/main/kotlin/taeyun/malanalter/ItemChecker.kt
@@ -1,9 +1,6 @@
 package taeyun.malanalter
 
 import kotlinx.coroutines.Job
-import taeyun.malanalter.alertitem.domain.ItemBidEntity
-import taeyun.malanalter.alertitem.dto.ItemBidInfo
-import taeyun.malanalter.alertitem.dto.RegisteredItem
 
 /**
  * 아이템 체크 및 알람 발송의 핵심 로직을 추상화한 인터페이스.
@@ -13,7 +10,4 @@ interface ItemChecker {
 
     /** 등록된 아이템을 확인하고 알람을 발송하는 Job을 실행한다. */
     fun checkItem(): Job
-
-    /** 특정 아이템의 비드를 조회하고 알람 대상 비드 목록을 반환한다. */
-    suspend fun requestItemBids(item: RegisteredItem, existBidList: List<ItemBidEntity>): List<ItemBidInfo>
 }
